+++ lib/ui/widgets/custom_logo.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:team_project/_core/constants/size.dart';
 
 class CustomLogo extends StatelessWidget {
   final String title;
@@ -7,18 +8,21 @@ class CustomLogo extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Image.asset(
-          "assets/carrot_image.png",
-          height: 100,
-          width: 100,
-        ),
-        Text(
-          title,
-          style: const TextStyle(fontSize: 40, fontWeight: FontWeight.bold),
-        ),
-      ],
+    return Padding(
+      padding: const EdgeInsets.only(top: xlargeGap, bottom: largeGap),
+      child: Column(
+        children: [
+          Image.asset(
+            "assets/carrot_image.png",
+            height: 100,
+            width: 100,
+          ),
+          Text(
+            title,
+            style: const TextStyle(fontSize: 40, fontWeight: FontWeight.bold),
+          ),
+        ],
+      ),
     );
   }
 }
